requirements.txt
biopython==1.84
build==1.2.2.post1
matplotlib==3.10.0
pandas==2.2.3
progress==1.6
scikit-learn==1.6.0
tensorflow==2.14.1
twine==6.0.1
umap-learn==0.5.7
